requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-environ==0.10.0
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
flake8==6.0.0
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.7.0
openapi-codec==1.3.2
phonenumbers==8.13.11
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.29.0
simplejson==3.19.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
